feat. 광고 쿠폰 시스템 개발 기능 운영 브랜치에 반영

+++ app/src/main/java/com/teumteumeat/teumteumeat/ui/screen/b2_quiz/QuizActivity.kt
@@ -123,7 +123,7 @@ class QuizActivity : BaseActivity() {
                             viewModel.loadQuizzes(documentId, GoalTypeUiState.fromString(goalType))
                         },
                         onCompleteQuiz = {
-                            val documentId = viewModel.documentId.toInt()
+                            val documentId = viewModel.documentId
 
                             viewModel.completeQuiz()
